+++ words/__init__.py
@@ -1,5 +1,10 @@
 """Words package."""
 
+__ALL__ = ["WordsError", "bp"]
+
 
 class WordsError(BaseException):
     """Words Exception."""
+
+
+bp = breakpoint