+++ words/cli/dm_cli.py
@@ -4,16 +4,18 @@ import click
 from rich import print as rprint
 from rich.table import Table
 
-from words.cli import console, err
+from words import WordsError, bp  # noqa
+from words.cli import console, err, pager
 from words.datamuse_api import DatamuseAPI
 from words.datamuse_options import DatamuseOptions
 from words.word_presenter import WordPresenter
 
-bp = breakpoint
-
 
 def _(**kwargs):
-    """Datamuse API wrapper."""
+    """Datamuse API wrapper.
+
+    Query the Datamuse API.
+    """
     has_required = filter(None, map(kwargs.get, DatamuseOptions.cli_required))
     if not tuple(has_required):
         err("No search option provided.")
@@ -26,19 +28,25 @@ def _(**kwargs):
         rprint(api.data)
         return
 
-    if api.long:
-        table = Table(*WordPresenter.headers())
-        for w in api.words:
-            word = WordPresenter(w)
-            table.add_row(*word.columns)
-        console.print(table)
-    else:
-        for word in api.words:
-            print(word)
+    with pager:
+        if api.long:
+            table = Table(*WordPresenter.headers())
+            for w in api.words:
+                word = WordPresenter(w)
+                table.add_row(*word.columns)
+            console.print(table)
+        else:
+            for word in api.words:
+                console.print(word)
 
 
-dm_cmd = click.Command("dm", callback=_, help=_.__doc__, params=[
-    *DatamuseOptions.cli_options,
-    click.Option(["--json"], is_flag=True, help="Export as raw JSON."),
-    click.Option(["--long"], is_flag=True, help="Print word list in long format."),
-])
+dm_cmd = click.Command(
+    "dm",
+    callback=_, help=_.__doc__,
+    epilog="Datamuse documentation: https://www.datamuse.com/api/",
+    params=[
+        *DatamuseOptions.cli_options,
+        click.Option(["--json"], is_flag=True, help="Export as raw JSON."),
+        click.Option(["--long"], is_flag=True, help="Print word list in long format."),
+    ],
+)