+++ words/datamuse_api.py
@@ -4,11 +4,10 @@ from functools import cached_property
 
 from datamuse import Datamuse as DatamuseClient
 
+from words import WordsError, bp  # noqa
 from words.datamuse_options import DatamuseOptions
 from words.object import Object
 from words.word import Word
-
-bp = breakpoint
 
 
 class DatamuseAPI(Object):